Groq_Meal_Plan/requirements.txt
groq
python-dotenv
streamlit
langchain
langchain-groq
fastapi
uvicorn
pydantic
pdfplumber
mysql-connector-python
